+++ src/pywwa/workflows/fake_afos_dump.py
@@ -29,15 +29,18 @@ NHC = {
     "URNT11": "REPNT1",
     "URNT12": "REPNT2",
     "URNT13": "REPNT3",
+    "UZNT13": "REPNT3",  # checking with NHC
     "URNT14": "REPNTS",
     "URPA10": "REPPA0",
     "URPA11": "REPPA1",
     "URPA12": "REPPA2",
     "URPA13": "REPPA3",
+    "UZPA13": "REPPA3",  # checking with NHC
     "URPN10": "REPPN0",
     "URPN11": "REPPN1",
     "URPN12": "REPPN2",
     "URPN13": "REPPN3",
+    "UZPN13": "REPPN3",  # checking with NHC
     "URPN14": "REPPNS",
 }