+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this library are documented in this file.
 
 ### Bug Fixes
 
+- Add fake afos entries for UZ(NT|PA|PN) for REPNT3.
 - Allow non-K sites through faked DSM AFOS creator.
 - Correct log setup for messages coming from pyIEM and pyWWA.
 - Populate all AFOS database fields for RR7 workflow.