+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 All notable changes to this library are documented in this file.
 
-## Unreleased Version
+## **1.7.1** (23 Sep 2025)
 
 ### API Changes